+++ app/(main)/(screens)/_layout.tsx
@@ -12,12 +12,7 @@ const ScreensLayout = () => {
       >
         <Stack.Screen name="notifications" />
         <Stack.Screen name="menu" />
-        <Stack.Screen name="brand-profile" />
-        <Stack.Screen name="profile" />
         <Stack.Screen name="CollaborationHistory" />
-        <Stack.Screen name="preferences" />
-        <Stack.Screen name="settings" />
-        <Stack.Screen name="billing" />
       </Stack>
     </AppLayout>
   );
