+++ app/(main)/(screens)/collaboration-details/[pageID].tsx
@@ -1,77 +1,46 @@
-import CollaborationPage from "@/components/collaboration/CollaborationDetails";
-import { useEffect, useState } from "react";
-import { doc, getDoc } from "firebase/firestore";
+import BottomSheetActions from "@/components/BottomSheetActions";
+import CollaborationDetails from "@/components/collaboration/collaboration-details";
+import BackButton from "@/components/ui/back-button/BackButton";
+import Colors from "@/constants/Colors";
+import AppLayout from "@/layouts/app-layout"
+import { useTheme } from "@react-navigation/native";
 import { useLocalSearchParams } from "expo-router";
-import { FirestoreDB } from "@/utils/firestore";
-import AppLayout from "@/layouts/app-layout";
-import { ActivityIndicator, Chip } from "react-native-paper";
-import { ICollaboration } from "@/shared-libs/firestore/trendly-pro/models/collaborations";
-import { View } from "@/components/theme/Themed";
-import { ScrollView, StyleSheet } from "react-native";
-
-interface Collaboration extends ICollaboration {
-  brandName: string;
-  paymentVerified: boolean;
-  brandDescription: string;
-}
+import { useState } from "react";
+import { Appbar } from "react-native-paper"
 
 const CollaborationDetailsScreen = () => {
+  const theme = useTheme();
+  const [isVisible, setIsVisible] = useState(false);
   const pageID = useLocalSearchParams().pageID;
-  const [collaboration, setCollaboration] = useState<Collaboration | undefined>(
-    undefined
-  ); // Explicit type
-  const [loading, setLoading] = useState(true);
-  const [activeTab, setActiveTab] = useState("Overview");
-  const fetchCollaboration = async () => {
-    if (!pageID) return;
-    try {
-      const collabRef = doc(FirestoreDB, "collaborations", pageID as string);
-      const snapshot = await getDoc(collabRef);
-      const data = snapshot.data() as ICollaboration;
-      if (!data) return;
-
-      const brandRef = doc(FirestoreDB, "brands", data.brandId);
-      const brandSnapshot = await getDoc(brandRef);
-      const brandData = brandSnapshot.data();
-
-      setCollaboration({
-        ...data,
-        brandName: brandData?.name || "Unknown Brand",
-        paymentVerified: brandData?.paymentMethodVerified || false,
-        brandDescription: brandData?.description || "",
-      });
-    } catch (e) {
-      console.error(e);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchCollaboration();
-  }, [pageID]);
-
-  if (loading)
-    return (
-      <AppLayout>
-        <ActivityIndicator />
-      </AppLayout>
-    );
-
-  if (!collaboration) return null;
 
   return (
     <AppLayout>
-      {activeTab === "Overview" && (
-        <CollaborationPage
-          collaborationDetail={collaboration}
-          pageID={pageID}
+      <Appbar.Header
+        statusBarHeight={0}
+        style={{
+          backgroundColor: Colors(theme).background,
+        }}
+      >
+        <BackButton />
+        <Appbar.Content
+          title="Collaboration Details"
+          color={Colors(theme).text}
         />
-      )}
+      </Appbar.Header>
+      <CollaborationDetails
+        pageID={pageID as string}
+      />
+      <BottomSheetActions
+        cardId={pageID as string}
+        cardType="influencerCard"
+        isVisible={isVisible}
+        snapPointsRange={["20%", "50%"]}
+        onClose={() => {
+          setIsVisible(false);
+        }}
+      />
     </AppLayout>
   );
 };
 
 export default CollaborationDetailsScreen;
-
-const styles = StyleSheet.create({});
