+++ components/drawer-layout/DrawerMenuItem.tsx
@@ -3,11 +3,12 @@ import { usePathname, useRouter } from "expo-router";
 import { Pressable, StyleSheet } from "react-native";
 import { Text, View } from "../theme/Themed";
 import Colors from "@/constants/Colors";
-import { FontAwesome } from "@expo/vector-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-native-fontawesome";
+import { IconProp } from "@fortawesome/fontawesome-svg-core";
 
 type Tab = {
   href: string;
-  icon: string;
+  icon: IconProp;
   label: string;
 };
 
@@ -38,7 +39,7 @@ const DrawerMenuItem: React.FC<DrawerMenuItemProps> = ({ tab }) => {
           justifyContent: "flex-start"
         }}
       >
-        <FontAwesome
+        {/* <FontAwesome
           name={tab.icon as any}
           color={
             tab.href.includes(pathname)
@@ -50,6 +51,18 @@ const DrawerMenuItem: React.FC<DrawerMenuItemProps> = ({ tab }) => {
             textAlign: "center",
           }}
           size={28}
+        /> */}
+        <FontAwesomeIcon
+          icon={tab.icon}
+          color={
+            tab.href.includes(pathname)
+              ? Colors(theme).white
+              : Colors(theme).gray100
+          }
+          size={28}
+          style={{
+            width: 32,
+          }}
         />
         <Text
           style={{
