+++ package.json
@@ -18,6 +18,11 @@
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
     "@firebase/rules-unit-testing": "^3.0.4",
+    "@fortawesome/fontawesome-svg-core": "^6.7.0",
+    "@fortawesome/free-brands-svg-icons": "^6.7.0",
+    "@fortawesome/free-regular-svg-icons": "^6.7.0",
+    "@fortawesome/free-solid-svg-icons": "^6.7.0",
+    "@fortawesome/react-native-fontawesome": "^0.3.2",
     "@gorhom/bottom-sheet": "^5.0.2",
     "@ptomasroos/react-native-multi-slider": "^2.2.2",
     "@react-native-async-storage/async-storage": "^1.24.0",
@@ -63,7 +68,7 @@
     "react-native-safe-area-context": "^4.10.5",
     "react-native-screens": "^3.31.1",
     "react-native-select-dropdown": "^4.0.1",
-    "react-native-svg": "15.2.0",
+    "react-native-svg": "^15.2.0",
     "react-native-swiper": "^1.6.0",
     "react-native-toast-message": "^2.2.1",
     "react-native-video": "^6.7.0",
