+++ utils/conversion.ts
@@ -1,0 +1,6 @@
+export const convertToKUnits = (num: number) => {
+  if (num >= 1000) {
+    return `${(num / 1000).toFixed(1)}k`;
+  }
+  return num;
+};
